refactor: move extractParameter() fun into Visitor class
