+++ processors/src/main/kotlin/com/alecarnevale/claymore/visitors/InterfaceAutoBindsVisitor.kt
@@ -6,9 +6,9 @@ import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.KSClassDeclaration
-import com.google.devtools.ksp.symbol.KSType
 import com.google.devtools.ksp.symbol.Modifier
 import com.squareup.kotlinpoet.ksp.writeTo
+import kotlin.reflect.KClass
 
 /**
  * This visitor check if provided implementation of [InterfaceAutoBinds] is a descendant of the annotated interface.
@@ -19,6 +19,8 @@ internal class InterfaceAutoBindsVisitor(
   override val resolver: Resolver,
   override val logger: KSPLogger
 ) : Visitor() {
+
+  override val kclass: KClass<*> = InterfaceAutoBinds::class
 
   override fun visitClassDeclaration(classDeclaration: KSClassDeclaration, data: Unit) {
     logger.info("$TAG visitClassDeclaration of $classDeclaration")
@@ -62,34 +64,6 @@ internal class InterfaceAutoBindsVisitor(
       aggregating = false,
       originatingKSFiles = dependencies
     )
-  }
-
-  private fun KSClassDeclaration.extractParameter(parameterName: String): KSClassDeclaration? {
-    // extract the KSType
-    val autobindsAnnotation =
-      annotations.firstOrNull { it.shortName.getShortName() == InterfaceAutoBinds::class.simpleName }
-    val parameterKsType =
-      autobindsAnnotation?.arguments?.firstOrNull { it.name?.getShortName() == parameterName }?.value as? KSType
-    if (parameterKsType == null) {
-      logger.error("$TAG parameter class not found for $parameterName")
-      return null
-    }
-
-    // extract the KSName
-    val parameterQualifiedName = parameterKsType.declaration.qualifiedName
-    if (parameterQualifiedName == null) {
-      logger.error("$TAG qualified name is null for $parameterName")
-      return null
-    }
-
-    // extract the KSClassDeclaration
-    val parameterDeclaration = resolver.getClassDeclarationByName(parameterQualifiedName)
-    if (parameterDeclaration == null) {
-      logger.error("$TAG implementation class not found for $parameterName")
-      return null
-    }
-
-    return parameterDeclaration
   }
 }
